feat: UI cleanup and theme improvements
- Move AI Recipe Creator button from navbar to recipes page
- Fix button colors to use dark brown Caramelatte theme instead of black
- Replace DebugAuthProvider with SimpleAuthProvider for better reliability
- Update theme documentation with current brown color palette
- Fix CSS custom properties conflict that was overriding DaisyUI theme
- Reorganize docs into auth/ and theme/ subdirectories
- Ensure all buttons use @ui/ components with consistent styling

Theme colors updated:
- Primary: #3d2817 (dark brown)
- Secondary: #6b4423 (medium brown)
- Accent: #8b5a2b (lighter brown)
- Neutral: #5d4037 (rich brown)

Resolves button color issues and improves overall UI consistency.
